Computation of Specific area and Plotting The Coordinates

Provide an approximation of the area of a Shape by using Monte Carlo Approach, with the function estimate_area(), we can simulate coordinates of x and y , which are uniformly distributed between 0 and 1 according to the number that we write in the function. At the end, we have an approximate area calculated by counting the number of points inside the shape divided by the total number of points.

estimate_area(B = 5000, seed = 10)

Arguments

B
A numeric (integer) used to denote the number of coordinates for simulation.

seed
A numeric used to control the seed of the random number generator used by this function.

Value

A list containing the following attributes:

Estimated_area

Estimated area of a shape

points

Coordinates of simulated points

Author

Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesca Darino

Examples

estimate_area(B = 500, seed = 100)
#> $estimated_area
#> [1] 0.586
#> 
#> $points
#>               x            y inside
#> 1   0.307766111 0.2202325612  FALSE
#> 2   0.257672501 0.7431595023   TRUE
#> 3   0.552322433 0.3059573120   TRUE
#> 4   0.056383150 0.4022088493  FALSE
#> 5   0.468549284 0.3935714429   TRUE
#> 6   0.483770735 0.4407564539   TRUE
#> 7   0.812402618 0.7728050761   TRUE
#> 8   0.370320537 0.2095690435  FALSE
#> 9   0.546558595 0.7481226882   TRUE
#> 10  0.170262051 0.5853551526   TRUE
#> 11  0.624996477 0.3244471955   TRUE
#> 12  0.882165518 0.3750474500  FALSE
#> 13  0.280353840 0.0269371939  FALSE
#> 14  0.398487902 0.4833348761   TRUE
#> 15  0.762551082 0.9700804388  FALSE
#> 16  0.669021712 0.0003950703  FALSE
#> 17  0.204612162 0.1057025485  FALSE
#> 18  0.357524853 0.4399432044   TRUE
#> 19  0.359475115 0.3122987410  FALSE
#> 20  0.690290528 0.9713328420  FALSE
#> 21  0.535811154 0.5852688691   TRUE
#> 22  0.710803845 0.7339286285   TRUE
#> 23  0.538348698 0.6635271511   TRUE
#> 24  0.748972227 0.7229845193   TRUE
#> 25  0.420101451 0.1110149436  FALSE
#> 26  0.171420213 0.2185792171  FALSE
#> 27  0.770301610 0.6225663018   TRUE
#> 28  0.881953588 0.9592804280  FALSE
#> 29  0.549096710 0.8123157280   TRUE
#> 30  0.277723757 0.0221628349  FALSE
#> 31  0.488305994 0.6825390456   TRUE
#> 32  0.928505075 0.8307875313  FALSE
#> 33  0.348691982 0.8078064425   TRUE
#> 34  0.954157708 0.8418451962  FALSE
#> 35  0.695274139 0.1876676735  FALSE
#> 36  0.889453539 0.0949448342  FALSE
#> 37  0.180407245 0.4488151143  FALSE
#> 38  0.629390850 0.6004804880   TRUE
#> 39  0.989564137 0.8635722811  FALSE
#> 40  0.130288870 0.8514558950  FALSE
#> 41  0.330660526 0.6977811693   TRUE
#> 42  0.865120547 0.6439721654   TRUE
#> 43  0.777584439 0.6546971288   TRUE
#> 44  0.827303450 0.8803548680  FALSE
#> 45  0.603324356 0.8569371656   TRUE
#> 46  0.491231822 0.8468295794   TRUE
#> 47  0.780358511 0.3265210520   TRUE
#> 48  0.884227027 0.5470013041   TRUE
#> 49  0.207713898 0.7343759690   TRUE
#> 50  0.307085897 0.8052955386   TRUE
#> 51  0.330529848 0.2064400080  FALSE
#> 52  0.198679070 0.8801616649   TRUE
#> 53  0.235694302 0.3960763614  FALSE
#> 54  0.274886660 0.9830069796  FALSE
#> 55  0.591321054 0.5614219203   TRUE
#> 56  0.253390654 0.9086144671   TRUE
#> 57  0.123487230 0.7151337676   TRUE
#> 58  0.229905887 0.2356364124  FALSE
#> 59  0.597575292 0.0039051792  FALSE
#> 60  0.211408556 0.4793805082   TRUE
#> 61  0.463701178 0.8721331256   TRUE
#> 62  0.647101194 0.2623272671   TRUE
#> 63  0.960573092 0.6634045555   TRUE
#> 64  0.676398172 0.8796447231   TRUE
#> 65  0.445148022 0.9115052556   TRUE
#> 66  0.357773783 0.3903184666   TRUE
#> 67  0.455731456 0.8651506912   TRUE
#> 68  0.445413977 0.5146185814   TRUE
#> 69  0.245092589 0.7411989463   TRUE
#> 70  0.694350712 0.9602066136   TRUE
#> 71  0.412237035 0.5697099713   TRUE
#> 72  0.327725869 0.3110995081  FALSE
#> 73  0.572564767 0.7433016279   TRUE
#> 74  0.966999084 0.4504729696  FALSE
#> 75  0.661779022 0.4738918107   TRUE
#> 76  0.624697717 0.4744165663   TRUE
#> 77  0.856653043 0.7407055742   TRUE
#> 78  0.774778888 0.1071925669  FALSE
#> 79  0.834027099 0.6421807767   TRUE
#> 80  0.091510279 0.4221840035  FALSE
#> 81  0.459525487 0.8005092293   TRUE
#> 82  0.599398156 0.6167882476   TRUE
#> 83  0.919721911 0.2493683256  FALSE
#> 84  0.982824077 0.6812233408  FALSE
#> 85  0.037802579 0.3477386320  FALSE
#> 86  0.577937401 0.8914637144   TRUE
#> 87  0.733314168 0.6572394262   TRUE
#> 88  0.248742403 0.9386654657  FALSE
#> 89  0.300736525 0.1341775106  FALSE
#> 90  0.733466702 0.9724600734  FALSE
#> 91  0.906954376 0.3507485189  FALSE
#> 92  0.209816768 0.2370002337  FALSE
#> 93  0.358137989 0.6156287836   TRUE
#> 94  0.448299144 0.7471337013   TRUE
#> 95  0.906426433 0.9635183825  FALSE
#> 96  0.389439296 0.9326846676   TRUE
#> 97  0.517459749 0.7921867422   TRUE
#> 98  0.125239088 0.3117346237  FALSE
#> 99  0.030145746 0.0934134405  FALSE
#> 100 0.771805495 0.2072938555  FALSE
#> 101 0.327415082 0.0317663355  FALSE
#> 102 0.389478693 0.5797054924   TRUE
#> 103 0.041052755 0.1542048361  FALSE
#> 104 0.361396634 0.1252705026  FALSE
#> 105 0.570978084 0.1479858113   TRUE
#> 106 0.684880242 0.9133426277   TRUE
#> 107 0.971116747 0.2574304065  FALSE
#> 108 0.701958807 0.5424652954   TRUE
#> 109 0.011545505 0.6490224628  FALSE
#> 110 0.535532129 0.1491943700   TRUE
#> 111 0.836571463 0.8489404500   TRUE
#> 112 0.806866742 0.2935050817  FALSE
#> 113 0.080932643 0.5521420008   TRUE
#> 114 0.238909302 0.8731348554   TRUE
#> 115 0.965989864 0.6469844421   TRUE
#> 116 0.037672326 0.3971833922  FALSE
#> 117 0.916406432 0.0295334503  FALSE
#> 118 0.726226578 0.3568591708   TRUE
#> 119 0.200776214 0.6985446722   TRUE
#> 120 0.840200746 0.2994483276  FALSE
#> 121 0.396662347 0.8048719901   TRUE
#> 122 0.392702259 0.7742791111   TRUE
#> 123 0.472556887 0.2111351776   TRUE
#> 124 0.583656915 0.3771634167   TRUE
#> 125 0.352387169 0.3113736284  FALSE
#> 126 0.028474013 0.8975737686  FALSE
#> 127 0.995087935 0.4164614563  FALSE
#> 128 0.957558459 0.8358839818  FALSE
#> 129 0.551651176 0.3214397281   TRUE
#> 130 0.101953957 0.1373967084  FALSE
#> 131 0.237915148 0.6177129149   TRUE
#> 132 0.860097665 0.6754165094   TRUE
#> 133 0.738261286 0.2724321040   TRUE
#> 134 0.497312265 0.1862069692   TRUE
#> 135 0.579921085 0.3303263714   TRUE
#> 136 0.016183385 0.1218199988  FALSE
#> 137 0.472129882 0.2355305520   TRUE
#> 138 0.042389729 0.9328466295  FALSE
#> 139 0.463154416 0.7825414138   TRUE
#> 140 0.629994768 0.4517651494   TRUE
#> 141 0.673249288 0.1108000707  FALSE
#> 142 0.087034671 0.8886778960  FALSE
#> 143 0.143583800 0.8135874579   TRUE
#> 144 0.908478499 0.1967867636  FALSE
#> 145 0.122532590 0.6002224234   TRUE
#> 146 0.728862792 0.0399579627  FALSE
#> 147 0.950377176 0.4737619837   TRUE
#> 148 0.043355915 0.7769649457  FALSE
#> 149 0.019599298 0.5798329108   TRUE
#> 150 0.199072460 0.3219917372  FALSE
#> 151 0.505086262 0.9933883986   TRUE
#> 152 0.926388763 0.4962693667   TRUE
#> 153 0.138401692 0.6814483758   TRUE
#> 154 0.169194576 0.7379226335   TRUE
#> 155 0.606627378 0.9076129526   TRUE
#> 156 0.815876693 0.4414791761   TRUE
#> 157 0.843381207 0.7481241769   TRUE
#> 158 0.788077194 0.7545178870   TRUE
#> 159 0.019020866 0.4453595262  FALSE
#> 160 0.698087497 0.3663760524   TRUE
#> 161 0.815093115 0.0790969988  FALSE
#> 162 0.568400061 0.2051951340   TRUE
#> 163 0.480062883 0.8850008196   TRUE
#> 164 0.161370118 0.9169217129  FALSE
#> 165 0.089256232 0.6175643548   TRUE
#> 166 0.162682375 0.3660410407  FALSE
#> 167 0.026728450 0.6758865851  FALSE
#> 168 0.709629476 0.4104196595   TRUE
#> 169 0.761018177 0.6693352878   TRUE
#> 170 0.857474779 0.9101494078  FALSE
#> 171 0.437265820 0.4148754235   TRUE
#> 172 0.417136447 0.1529516319  FALSE
#> 173 0.585649490 0.3774737197   TRUE
#> 174 0.824967581 0.4657902545   TRUE
#> 175 0.793139227 0.2442707589  FALSE
#> 176 0.326152570 0.1182193570  FALSE
#> 177 0.957931943 0.0082823546  FALSE
#> 178 0.653019610 0.6508688976   TRUE
#> 179 0.458675819 0.0565655837  FALSE
#> 180 0.606021883 0.1301092303   TRUE
#> 181 0.288721937 0.9606898576  FALSE
#> 182 0.717830844 0.2595715576   TRUE
#> 183 0.923397443 0.0090666011  FALSE
#> 184 0.674939486 0.9928706293  FALSE
#> 185 0.185939992 0.6529038022   TRUE
#> 186 0.348067163 0.7244848204   TRUE
#> 187 0.123519531 0.7142819883   TRUE
#> 188 0.108116219 0.8250241783  FALSE
#> 189 0.297953200 0.1482449211  FALSE
#> 190 0.838199901 0.6750034729   TRUE
#> 191 0.992771094 0.8468728585  FALSE
#> 192 0.437134409 0.9925849929   TRUE
#> 193 0.202564571 0.3040931860  FALSE
#> 194 0.963871501 0.1453948889  FALSE
#> 195 0.660387112 0.5190369436   TRUE
#> 196 0.298116829 0.8321523182   TRUE
#> 197 0.119262186 0.9181715080  FALSE
#> 198 0.599751601 0.3481038627   TRUE
#> 199 0.120190553 0.3326893374  FALSE
#> 200 0.788432016 0.8388670043   TRUE
#> 201 0.369596071 0.4462395904   TRUE
#> 202 0.956322836 0.6945935639   TRUE
#> 203 0.913576656 0.9028108264  FALSE
#> 204 0.823336344 0.0294742039  FALSE
#> 205 0.319482156 0.4718077183   TRUE
#> 206 0.877700326 0.4088902143   TRUE
#> 207 0.800350998 0.3369402727   TRUE
#> 208 0.611356641 0.6632755490   TRUE
#> 209 0.072421137 0.8731358480  FALSE
#> 210 0.421587710 0.0867594131  FALSE
#> 211 0.344465601 0.3346435984  FALSE
#> 212 0.751735687 0.9337488771  FALSE
#> 213 0.218751320 0.9185929906  FALSE
#> 214 0.292133973 0.6867532160   TRUE
#> 215 0.355953365 0.3835435782   TRUE
#> 216 0.631961924 0.0089821673  FALSE
#> 217 0.892531067 0.7199387688   TRUE
#> 218 0.744032781 0.4809266145   TRUE
#> 219 0.457222422 0.0037227867  FALSE
#> 220 0.036095444 0.5721116830   TRUE
#> 221 0.568514518 0.0806550384   TRUE
#> 222 0.439015890 0.3798907280   TRUE
#> 223 0.600484454 0.9995236769  FALSE
#> 224 0.951382762 0.0974674660  FALSE
#> 225 0.269431312 0.8042160822   TRUE
#> 226 0.658838391 0.0842295843  FALSE
#> 227 0.076471213 0.8771823877  FALSE
#> 228 0.071285208 0.9825428384  FALSE
#> 229 0.370331520 0.6098533694   TRUE
#> 230 0.297102200 0.6848330807   TRUE
#> 231 0.551078272 0.4945951006   TRUE
#> 232 0.369890589 0.2759890214  FALSE
#> 233 0.845689532 0.0037211182  FALSE
#> 234 0.620790703 0.6644262737   TRUE
#> 235 0.399140022 0.5482790163   TRUE
#> 236 0.299661528 0.5960336337   TRUE
#> 237 0.381119837 0.3083463048  FALSE
#> 238 0.701134936 0.8972893313   TRUE
#> 239 0.946865291 0.9161570750  FALSE
#> 240 0.772828321 0.9716905395  FALSE
#> 241 0.219550162 0.3423762012  FALSE
#> 242 0.716110362 0.9284288513   TRUE
#> 243 0.664217965 0.9186325045   TRUE
#> 244 0.706250201 0.6914463339   TRUE
#> 245 0.279627993 0.7699238688   TRUE
#> 246 0.712250495 0.6021966012   TRUE
#> 247 0.660942108 0.7693214894   TRUE
#> 248 0.041012689 0.0979830830  FALSE
#> 249 0.061141495 0.2317523940  FALSE
#> 250 0.279761693 0.7068353218   TRUE
#> 251 0.301967717 0.0288219615  FALSE
#> 252 0.956610830 0.6311423669   TRUE
#> 253 0.389818719 0.6721577807   TRUE
#> 254 0.371788787 0.4528857339   TRUE
#> 255 0.843142486 0.6758988027   TRUE
#> 256 0.847235000 0.0093933111  FALSE
#> 257 0.319331153 0.0634803819  FALSE
#> 258 0.132506046 0.2792331835  FALSE
#> 259 0.617109120 0.4678869434   TRUE
#> 260 0.791304509 0.7764572625   TRUE
#> 261 0.338048706 0.3669242659  FALSE
#> 262 0.905447597 0.9007200717  FALSE
#> 263 0.197556710 0.5214052640   TRUE
#> 264 0.794085179 0.2896093260  FALSE
#> 265 0.754602895 0.2445833343  FALSE
#> 266 0.911394847 0.5891333013   TRUE
#> 267 0.322687705 0.3807563642  FALSE
#> 268 0.086169473 0.7054133334   TRUE
#> 269 0.911200454 0.1160407732  FALSE
#> 270 0.955173825 0.5984548617   TRUE
#> 271 0.671142944 0.8460731402   TRUE
#> 272 0.744982201 0.3508640204   TRUE
#> 273 0.440016976 0.3161670289   TRUE
#> 274 0.115039418 0.7801960679   TRUE
#> 275 0.675642782 0.2547899224   TRUE
#> 276 0.731559517 0.8402001492   TRUE
#> 277 0.483984903 0.7513929436   TRUE
#> 278 0.171048038 0.3087822665  FALSE
#> 279 0.675848535 0.4580469287   TRUE
#> 280 0.262989946 0.4148980144  FALSE
#> 281 0.341480829 0.3250160257  FALSE
#> 282 0.210245543 0.1356165456  FALSE
#> 283 0.016319594 0.9101109649  FALSE
#> 284 0.376762877 0.1503848655  FALSE
#> 285 0.562307402 0.8379229447   TRUE
#> 286 0.679984608 0.5442733990   TRUE
#> 287 0.745388770 0.1159527488  FALSE
#> 288 0.950072722 0.0442293435  FALSE
#> 289 0.163090714 0.2826934732  FALSE
#> 290 0.324756902 0.8812637459   TRUE
#> 291 0.132716035 0.8369479515   TRUE
#> 292 0.637297439 0.0577618242  FALSE
#> 293 0.330929611 0.0487210115  FALSE
#> 294 0.649079573 0.3166363316   TRUE
#> 295 0.302888359 0.1314321244  FALSE
#> 296 0.071233266 0.0356124358  FALSE
#> 297 0.662390469 0.1208346796  FALSE
#> 298 0.759603562 0.5678826217   TRUE
#> 299 0.553360169 0.1273408108   TRUE
#> 300 0.539272098 0.7580523530   TRUE
#> 301 0.849592320 0.0994260872  FALSE
#> 302 0.653225959 0.2299340756   TRUE
#> 303 0.950885756 0.4436262054  FALSE
#> 304 0.617223048 0.5157049047   TRUE
#> 305 0.492840634 0.9248942458   TRUE
#> 306 0.976006565 0.0444568363  FALSE
#> 307 0.490345216 0.1491565800   TRUE
#> 308 0.655172303 0.8481738840   TRUE
#> 309 0.598801789 0.1529559076   TRUE
#> 310 0.947569016 0.8249667988  FALSE
#> 311 0.368011507 0.8631921196   TRUE
#> 312 0.878187005 0.6543586599   TRUE
#> 313 0.454875068 0.0220086016  FALSE
#> 314 0.496467015 0.9979430493   TRUE
#> 315 0.460615615 0.1476474060  FALSE
#> 316 0.617908197 0.0825577057  FALSE
#> 317 0.604143492 0.7976932214   TRUE
#> 318 0.785879603 0.3440654075   TRUE
#> 319 0.555268475 0.7286060092   TRUE
#> 320 0.768978339 0.3410495843   TRUE
#> 321 0.404285673 0.8608303124   TRUE
#> 322 0.510719810 0.4072234400   TRUE
#> 323 0.523536443 0.5685604715   TRUE
#> 324 0.992799208 0.9015503402  FALSE
#> 325 0.429647421 0.9352645874   TRUE
#> 326 0.996011255 0.8472000654  FALSE
#> 327 0.786600247 0.4140229521   TRUE
#> 328 0.515910635 0.3474344276   TRUE
#> 329 0.502687966 0.4568983915   TRUE
#> 330 0.911637484 0.9037745774  FALSE
#> 331 0.264415897 0.3602777214  FALSE
#> 332 0.173856868 0.5791492232   TRUE
#> 333 0.400331255 0.6281828401   TRUE
#> 334 0.538357384 0.7134869962   TRUE
#> 335 0.244964375 0.0220602453  FALSE
#> 336 0.375850958 0.4969065846   TRUE
#> 337 0.580253547 0.4405028000   TRUE
#> 338 0.208955312 0.5577421037   TRUE
#> 339 0.801330007 0.1756188539  FALSE
#> 340 0.638802415 0.5635869007   TRUE
#> 341 0.736330756 0.6096523490   TRUE
#> 342 0.438812408 0.6390128997   TRUE
#> 343 0.579812383 0.7929890070   TRUE
#> 344 0.256044009 0.3464461036  FALSE
#> 345 0.463718232 0.3335031627   TRUE
#> 346 0.168614859 0.9874712862  FALSE
#> 347 0.613894533 0.7164507983   TRUE
#> 348 0.958102621 0.4110368493  FALSE
#> 349 0.478194729 0.5702433558   TRUE
#> 350 0.751159331 0.1049496075  FALSE
#> 351 0.020583217 0.8672685213  FALSE
#> 352 0.170969027 0.4778781414   TRUE
#> 353 0.639966496 0.2785666285   TRUE
#> 354 0.164978485 0.4630207387  FALSE
#> 355 0.354722775 0.0410848528  FALSE
#> 356 0.186426474 0.8743469159   TRUE
#> 357 0.897656161 0.7433382492   TRUE
#> 358 0.237783856 0.0726328203  FALSE
#> 359 0.984943485 0.4259650176  FALSE
#> 360 0.021330737 0.5157740989   TRUE
#> 361 0.107538965 0.7387045722   TRUE
#> 362 0.244826828 0.2680203572  FALSE
#> 363 0.722362406 0.4872992872   TRUE
#> 364 0.032734943 0.9623953735  FALSE
#> 365 0.549349994 0.8418851502   TRUE
#> 366 0.682808681 0.2681082212   TRUE
#> 367 0.300240897 0.2607396969  FALSE
#> 368 0.388774461 0.9133851803   TRUE
#> 369 0.732446152 0.5991072529   TRUE
#> 370 0.962496418 0.0636683633  FALSE
#> 371 0.760596170 0.7173663327   TRUE
#> 372 0.581673162 0.7421439788   TRUE
#> 373 0.462872993 0.2627069724   TRUE
#> 374 0.356309620 0.5300226500   TRUE
#> 375 0.383921816 0.8043558830   TRUE
#> 376 0.205781032 0.7278188856   TRUE
#> 377 0.138780371 0.8653947313  FALSE
#> 378 0.389082144 0.0463959803  FALSE
#> 379 0.266046223 0.0080812450  FALSE
#> 380 0.703296282 0.9393897916   TRUE
#> 381 0.407878167 0.2416624320  FALSE
#> 382 0.265564194 0.8997069411   TRUE
#> 383 0.400977855 0.5113481141   TRUE
#> 384 0.197313132 0.6969351373   TRUE
#> 385 0.829931688 0.9570553503  FALSE
#> 386 0.527987730 0.0541673789   TRUE
#> 387 0.395130190 0.1295977051  FALSE
#> 388 0.573966566 0.7682044986   TRUE
#> 389 0.970972070 0.8613149433  FALSE
#> 390 0.650149828 0.7278726187   TRUE
#> 391 0.333601159 0.9042041593   TRUE
#> 392 0.331447429 0.7415114422   TRUE
#> 393 0.942434883 0.1615115390  FALSE
#> 394 0.381112495 0.4824496026   TRUE
#> 395 0.564323923 0.4100493968   TRUE
#> 396 0.511515689 0.4659072638   TRUE
#> 397 0.138860555 0.4434991970  FALSE
#> 398 0.239789413 0.0265242036  FALSE
#> 399 0.718009098 0.6539522654   TRUE
#> 400 0.297223778 0.3905836442  FALSE
#> 401 0.511237420 0.0262239345   TRUE
#> 402 0.277710656 0.2910003348  FALSE
#> 403 0.360656928 0.0004488295  FALSE
#> 404 0.437527851 0.6454668748   TRUE
#> 405 0.803066731 0.8269246418   TRUE
#> 406 0.520609729 0.7549124290   TRUE
#> 407 0.696152071 0.5148178525   TRUE
#> 408 0.847836898 0.6195992520   TRUE
#> 409 0.845709252 0.3979892961   TRUE
#> 410 0.391875740 0.4305372024   TRUE
#> 411 0.153513757 0.6347250536   TRUE
#> 412 0.639425793 0.4797932676   TRUE
#> 413 0.287171022 0.5593633349   TRUE
#> 414 0.928337881 0.5097432989   TRUE
#> 415 0.155636119 0.8265101276   TRUE
#> 416 0.963841664 0.5600976583   TRUE
#> 417 0.001260480 0.0132807584  FALSE
#> 418 0.707526205 0.4755937520   TRUE
#> 419 0.630187602 0.0364418153  FALSE
#> 420 0.773078045 0.2448342000  FALSE
#> 421 0.892606881 0.9546872040  FALSE
#> 422 0.511204083 0.2696251650   TRUE
#> 423 0.749001073 0.0472132596  FALSE
#> 424 0.926379797 0.9617045652  FALSE
#> 425 0.091753519 0.0411765689  FALSE
#> 426 0.496655486 0.5685715808   TRUE
#> 427 0.197501260 0.4893754795   TRUE
#> 428 0.994132901 0.8156825237  FALSE
#> 429 0.036821042 0.5262332316   TRUE
#> 430 0.217563070 0.8368379199   TRUE
#> 431 0.890936143 0.0718262950  FALSE
#> 432 0.118936464 0.7455581927   TRUE
#> 433 0.369038994 0.3374601963   TRUE
#> 434 0.162376430 0.6204041298   TRUE
#> 435 0.166691519 0.7179785939   TRUE
#> 436 0.972504712 0.9496248781  FALSE
#> 437 0.810399744 0.2719512167  FALSE
#> 438 0.953327630 0.9718085302  FALSE
#> 439 0.399910527 0.4724074309   TRUE
#> 440 0.852247320 0.7976263147   TRUE
#> 441 0.064522857 0.9088888676  FALSE
#> 442 0.302528767 0.9906902050  FALSE
#> 443 0.490472868 0.2213202249   TRUE
#> 444 0.646261239 0.1924467573   TRUE
#> 445 0.510634326 0.7099349364   TRUE
#> 446 0.797259235 0.8457011958   TRUE
#> 447 0.565008945 0.0437571695  FALSE
#> 448 0.353523986 0.7283248159   TRUE
#> 449 0.659099944 0.2991912132   TRUE
#> 450 0.239274576 0.2576841901  FALSE
#> 451 0.265210788 0.9804799608  FALSE
#> 452 0.518268851 0.8653444445   TRUE
#> 453 0.737910869 0.6641179975   TRUE
#> 454 0.779610906 0.8688727603   TRUE
#> 455 0.114694396 0.8434061243  FALSE
#> 456 0.612373349 0.9666821721   TRUE
#> 457 0.910936175 0.9132655612  FALSE
#> 458 0.634021459 0.2836070871   TRUE
#> 459 0.274664033 0.4578090804   TRUE
#> 460 0.352835338 0.8689027121   TRUE
#> 461 0.670829357 0.5165516310   TRUE
#> 462 0.906711419 0.2658771435  FALSE
#> 463 0.730235832 0.8686213219   TRUE
#> 464 0.919766989 0.6901659735   TRUE
#> 465 0.382433257 0.5945103467   TRUE
#> 466 0.854140538 0.9235723831  FALSE
#> 467 0.054734090 0.1493828474  FALSE
#> 468 0.253986029 0.8064511195   TRUE
#> 469 0.367128998 0.3114460232  FALSE
#> 470 0.360570408 0.9208744697   TRUE
#> 471 0.280516531 0.6495810472   TRUE
#> 472 0.388692510 0.1783565462  FALSE
#> 473 0.348134442 0.4413796300   TRUE
#> 474 0.278923067 0.7546392297   TRUE
#> 475 0.801088135 0.2576926639  FALSE
#> 476 0.995013841 0.9989377656  FALSE
#> 477 0.753241962 0.5224501409   TRUE
#> 478 0.603342928 0.9981128222  FALSE
#> 479 0.593948180 0.4026084305   TRUE
#> 480 0.582290288 0.5843024231   TRUE
#> 481 0.238659285 0.8918118891   TRUE
#> 482 0.454284382 0.8534271207   TRUE
#> 483 0.995516626 0.4292331270  FALSE
#> 484 0.082745983 0.2787725986  FALSE
#> 485 0.051906005 0.8918579654  FALSE
#> 486 0.493797213 0.9873077427   TRUE
#> 487 0.053993596 0.1395892045  FALSE
#> 488 0.950448573 0.5403982888   TRUE
#> 489 0.633191261 0.3105859275   TRUE
#> 490 0.269937081 0.8520803752   TRUE
#> 491 0.996812936 0.0435301224  FALSE
#> 492 0.358804111 0.3721722844   TRUE
#> 493 0.371762078 0.5159765771   TRUE
#> 494 0.330226399 0.7677630151   TRUE
#> 495 0.104395513 0.2873498579  FALSE
#> 496 0.579901429 0.8166487757   TRUE
#> 497 0.333089466 0.0052734755  FALSE
#> 498 0.003777579 0.0223096041  FALSE
#> 499 0.994596109 0.2429706121  FALSE
#> 500 0.369234861 0.8313589729   TRUE
#> 
#> attr(,"class")
#> [1] "area"